string security fixes

+++ src/random.c
@@ -11,5 +11,5 @@ void a_Random_Seed(void)
 
 uint32_t a_Random_Get32(void)
 {
-    return (uint32_t)(rand() % UINT32_MAX);
+    return rand() % UINT32_MAX;
 }
